localmovies-rag/requirements.txt
# Core RAG dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.22
sentence-transformers>=2.2.2

# API and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0
webdriver-manager>=4.0.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Web framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Environment and config
python-dotenv>=1.0.0
pydantic>=2.5.0

# Additional utilities
tqdm>=4.66.0
python-dateutil>=2.8.2 